+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Renamed the plugin to Again and the replay command to Previous command.
+- Added a configurable desktop status-bar display for the previous command.
+
 ## [1.0.0] - 2026-04-28
 
 - Initial release.